% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmsmanInteraction.R
\name{CreatehelmsmanOutput}
\alias{CreatehelmsmanOutput}
\title{Prepare input file for helmsman from a
helmsman formatted catalog file.}
\usage{
CreatehelmsmanOutput(
  catalog,
  out.dir = paste0(dirname(catalog), "/ExtrAttr/helmsman.results"),
  overwrite = FALSE
)
}
\arguments{
\item{catalog}{a catalog in ICAMS format. It can be
a .csv file, or a matrix or data.frame.
Usually, it refers to \code{"ground.truth.syn.catalog.csv"}.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exists. Usually, the \code{out.dir} will
be a \code{helmsman.results} folder directly under the
folder storing \code{catalog}.}

\item{overwrite}{If TRUE, overwrite existing output}
}
\value{
\code{invisible(catMatrix)},
original catalog in helmsman format
}
\description{
Prepare input file for helmsman from a
helmsman formatted catalog file.
}
\details{
Creates folder named \code{helmsman.results} containing catalogs
in helmsman-formatted catalogs: Rows are signatures;
the first column is the name of the mutation type, while the remaining
columns are samples (tumors).
These helmsman-formated catalogs will the input when running helmsman program
later on Python platform.
}
