% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcsmInteraction.R
\name{CreatetcsmOutput}
\alias{CreatetcsmOutput}
\title{Prepare input file for tcsm from a
tcsm formatted catalog file.}
\usage{
CreatetcsmOutput(
  catalog,
  out.dir = paste0(dirname(catalog), "/ExtrAttr/tcsm.results"),
  overwrite = FALSE
)
}
\arguments{
\item{catalog}{a catalog in ICAMS format. It can be
a .csv file, or a matrix or data.frame.
Usually, it refers to \code{"ground.truth.syn.catalog.csv"}.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exists. Usually, the \code{out.dir} will
be a \code{tcsm.results} folder directly under the
folder storing \code{catalog}.}

\item{overwrite}{If TRUE, overwrite existing output}
}
\value{
\code{invisible(catMatrix)},
original catalog in tcsm format
}
\description{
Prepare input file for tcsm from a
tcsm formatted catalog file.
}
\details{
Creates folder named \code{tcsm.results} containing catalogs
in tcsm-formatted catalogs: Rows are signatures;
the first column is the name of the mutation type, while the remaining
columns are samples (tumors).
These tcsm-formated catalogs will the input when running tcsm program
later on Python platform.
}
