% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmsmanInteraction.R
\name{helmsmanCatalog2ICAMS}
\alias{helmsmanCatalog2ICAMS}
\title{Convert Catalogs from helmsman format to ICAMS format}
\usage{
helmsmanCatalog2ICAMS(cat)
}
\arguments{
\item{cat}{Input catalog, can be a tab-delimited file
or matrix in helmsman format.}
}
\value{
a catalog matrix in ICAMS format.
}
\description{
Convert Catalogs from helmsman format to ICAMS format
}
