% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMuinteraction.R
\name{ICAMSCatalog2EMu}
\alias{ICAMSCatalog2EMu}
\title{Convert Catalogs from ICAMS format to EMu format}
\usage{
ICAMSCatalog2EMu(catalog)
}
\arguments{
\item{catalog}{A catalog matrix in ICAMS format. (SNS only!)}
}
\value{
a catalog matrix in EMu format.
}
\description{
Convert Catalogs from ICAMS format to EMu format
}
