% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmsmanInteraction.R
\name{ICAMSCatalog2helmsman}
\alias{ICAMSCatalog2helmsman}
\title{Convert Catalogs from ICAMS format to helmsman format}
\usage{
ICAMSCatalog2helmsman(catalog, type = "spectra")
}
\arguments{
\item{catalog}{A catalog matrix in ICAMS format. (SNS only!)}

\item{type}{Whether it is a spectra catalog ("spectra") or
a signature catalog ("signature").}
}
\value{
a catalog matrix in helmsman format.
}
\description{
Convert Catalogs from ICAMS format to helmsman format
}
