% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcsmInteraction.R
\name{ICAMSCatalog2tcsm}
\alias{ICAMSCatalog2tcsm}
\title{Convert Catalogs from ICAMS format to tcsm format}
\usage{
ICAMSCatalog2tcsm(catalog)
}
\arguments{
\item{catalog}{A catalog matrix in ICAMS format. (SNS only!)}
}
\value{
a catalog matrix in tcsm format.
}
\description{
Convert Catalogs from ICAMS format to tcsm format
}
