% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runmaftools.R
\name{Installmaftools}
\alias{Installmaftools}
\title{Install maftools from Bioconductor,
and its dependent package, NMF.}
\usage{
Installmaftools()
}
\description{
Install maftools from Bioconductor,
and its dependent package, NMF.
}
\keyword{internal}
