% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runsigminer.R
\name{Installsigminer}
\alias{Installsigminer}
\title{Install sigminer from Bioconductor,
also installing its dependent package, NMF.}
\usage{
Installsigminer()
}
\description{
Install sigminer from Bioconductor,
also installing its dependent package, NMF.
}
\keyword{internal}
