% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSomaticSignatures.R
\name{InstallSomaticSignatures}
\alias{InstallSomaticSignatures}
\title{Install SomaticSignatures from Bioconductor,
also installing its dependent package, NMF.}
\usage{
InstallSomaticSignatures()
}
\description{
Install SomaticSignatures from Bioconductor,
also installing its dependent package, NMF.
}
\keyword{internal}
