% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSparseSignatures.R
\name{InstallSparseSignatures}
\alias{InstallSparseSignatures}
\title{Install SparseSignatures from Bioconductor}
\usage{
InstallSparseSignatures()
}
\description{
Install SparseSignatures from Bioconductor
}
\keyword{internal}
