% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtcsm.R
\name{make.heldout.obj}
\alias{make.heldout.obj}
\title{tcsm's internal function to calculate likelihood for each \code{K}
using heldout method.}
\usage{
make.heldout.obj(train.mc.data, test.mc.data)
}
\arguments{
\item{train.mc.data}{Transposed \code{\link[ICAMS]{ICAMS}} spectra catalog
containing mutations of training samples.}

\item{test.mc.data}{Transposed \code{\link[ICAMS]{ICAMS}} spectra catalog
containing mutations of testing samples.}
}
\value{
heldout, which is a named list
  $missing
    $index
    $docs - list where each element is a list containing the mutations for a given sample
  $documents - list where each element is a list containing the mutations for a given sample
  $vocab - list the vocabularies, which refers to mutation types.
}
\description{
tcsm advises to use "heldout" method to obtain the likelihood of number of
mutational signatures, \code{K}.
That is, to split the whole spectra catalogs dataset into 2 parts:
}
\details{
80% of spectra catalogs should be allocated into \code{train.mc.data}
for training \code{\link[stm]{stm}} model,
20% of spectra catalogs should be allocated into \code{test.mc.data}
for estmating likelihood for given K.
}
\keyword{internal}
