% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmsmanInteraction.R
\name{ReadCathelmsman}
\alias{ReadCathelmsman}
\title{Read Catalog files in helmsman format}
\usage{
ReadCathelmsman(cat)
}
\arguments{
\item{cat}{Input catalog, can be a tab-delimited text
file in helmsman format, or a matrix/data.frame object.}
}
\value{
a catalog matrix in ICAMS format.
}
\description{
Read Catalog files in helmsman format
}
