% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMuinteraction.R
\name{ReadEMuCatalog}
\alias{ReadEMuCatalog}
\title{Read Catalog files in EMu format.}
\usage{
ReadEMuCatalog(
  cat,
  mutTypes,
  sigOrSampleNames,
  region = "unknown",
  catalog.type = "counts.signature"
)
}
\arguments{
\item{cat}{A tab-delimited catalog text file in EMu format;
or a EMu formatted matrix or data.frame.}

\item{mutTypes}{Types of mutations. They are usually from an
\code{ICAMS:::catalog.row.header} object.}

\item{sigOrSampleNames}{If input file is a counts signature file
\code{(catalog.type == "counts.signature")},
signature names should be provided.

If input file is a counts spectra file
\code{(catalog.type == "counts")},
names of samples should be provided.}

\item{region}{Catalog region. Can be a specific genomic
or exomic region, or "unknown".
Default: "unknown"}

\item{catalog.type}{Is the catalog a signature catalog,
or a spectrum catalog?
Default: "counts"}
}
\value{
a catalog matrix in ICAMS format.
}
\description{
Read Catalog files in EMu format.
}
