% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMuinteraction.R
\name{ReadEMuExposure}
\alias{ReadEMuExposure}
\title{Read Exposure files in EMu format.}
\usage{
ReadEMuExposure(exposure)
}
\arguments{
\item{exposure}{Exposure file generated by EMu.
Usually, it is called "W_components.txt".

Return ICAMS/SynSigEval formatted exposure matrix.}
}
\description{
Read Exposure files in EMu format.
}
