% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMuinteraction.R
\name{ReadEMuExposureFile}
\alias{ReadEMuExposureFile}
\title{Read Exposure files in EMu format.}
\usage{
ReadEMuExposureFile(exposureFile, sigNames, sampleNames)
}
\arguments{
\item{exposureFile}{Exposure file generated by EMu.
Usually, it is called "W_components.txt".}

\item{sigNames}{Names of signatures. These will be
served as the rownames of the exposure matrix.}

\item{sampleNames}{Names of samples in exposure file.

Return ICAMS/SynSigEval formatted exposure matrix.}
}
\description{
Read Exposure files in EMu format.
}
