% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalMuSigInteraction.R
\name{ReadExposureMM}
\alias{ReadExposureMM}
\title{Read Catalog files in MM format}
\usage{
ReadExposureMM(exposureFile)
}
\arguments{
\item{exposureFile}{Input exposure file, can be a tab-delimited
text file in MultiModalMuSig format.}
}
\value{
a exposure matrix in ICAMS format.
}
\description{
Read Catalog files in MM format
}
