% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmsmanInteraction.R
\name{ReadhelmsmanExposure}
\alias{ReadhelmsmanExposure}
\title{Read Exposure files in helmsman format.}
\usage{
ReadhelmsmanExposure(exposure, check.names = TRUE)
}
\arguments{
\item{exposure}{Exposure file generated by helmsman.
Usually, it is called "W_components.txt".}

\item{check.names}{logical. If \code{TRUE} then the names of the
variables in the data frame are checked to ensure that they are
syntactically valid variable names. If necessary they are adjusted
(by \code{\link[base]{make.names}}) so that they are, and also to
ensure that there are no duplicates.

Return ICAMS/SynSigEval formatted exposure matrix.}
}
\description{
Read Exposure files in helmsman format.
}
