% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigProfilerInteraction.R
\name{ReadSigProfilerSigDBS78}
\alias{ReadSigProfilerSigDBS78}
\title{Read a file containing DBS78 signatures extracted by SigProfiler/Python}
\usage{
ReadSigProfilerSigDBS78(file)
}
\arguments{
\item{file}{The name of the file to read.}
}
\value{
The corresponding signature matrix in standard internal
representation.
}
\description{
Read a file containing DBS78 signatures extracted by SigProfiler/Python
}
