% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigProfilerInteraction.R
\name{ReadSigProfilerSigSBS96}
\alias{ReadSigProfilerSigSBS96}
\title{Read a file containing SBS96 signatures extracted by SigProfiler/Python}
\usage{
ReadSigProfilerSigSBS96(file)
}
\arguments{
\item{file}{The name of the file to read.}
}
\value{
The corresponding signature matrix in standard internal
representation.
}
\description{
Read a file containing SBS96 signatures extracted by SigProfiler/Python
}
