% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcsmInteraction.R
\name{ReadtcsmExposure}
\alias{ReadtcsmExposure}
\title{Read Exposure files in tcsm format.}
\usage{
ReadtcsmExposure(exposure)
}
\arguments{
\item{exposure}{Exposure file generated by tcsm.
Usually, it is called "W_components.txt".

Return ICAMS/SynSigEval formatted exposure matrix.}
}
\description{
Read Exposure files in tcsm format.
}
