% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runhdp.R
\name{Runhdp}
\alias{Runhdp}
\title{Run hdp extraction and attribution on a spectra catalog file}
\usage{
Runhdp(
  input.catalog,
  out.dir,
  seedNumber = 1,
  K = NULL,
  K.range = NULL,
  multi.types = FALSE,
  remove.noise = FALSE,
  test.only = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input.catalog}{File containing a spectra catalog
in \code{\link[ICAMS]{ICAMS}} format.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{seedNumber}{Specify the pseudo-random seed number
used to run hdp. Setting seed can make the
attribution of hdp repeatable.
Default: 1.}

\item{K}{Suggested initial value of the number of
signatures, passed to \code{\link[hdp]dp_activate} as
\code{initcc}.

\code{K.range} Deprecated. Top value is interpreted as parameter \code{K}.}

\item{multi.types}{A logical scalar (\code{TRUE} or \code{FALSE}) or
a character vector.
If \code{FALSE}, hdp will regard all input spectra as one tumor type,
and will allocate them to one single dirichlet process node.

If \code{TRUE}, hdp will infer tumor types based on the string before "::" in their names.
e.g. Tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If it is a character vector, it should be a vector of case-sensitive tumor
types.
e.g. c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")

Default: FALSE}

\item{remove.noise}{Whether to remove noise signature "hdp.0"? In normal cases scenarios,
only few mutations will be assigned to noise signature.

For result visualization and assessment of \code{hdp} package, select \code{TRUE};
for diagnostic purposes, select \code{FALSE}.

Default: \code{FALSE}}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.
Default: FALSE}

\item{overwrite}{If TRUE, overwrite existing output.
Default: FALSE}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}
}
\value{
The attributed exposure of \code{hdp}, invisibly.
}
\description{
WARNING: hdp can only do exposure attribution
using SBS96 spectra catalog and signature catalog!
}
\details{
Creates several
 files in \code{out.dir}. These are:
 TODO(Steve): list the files

 TODO(Wuyang)
}
