% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runhdp2.R
\name{Runhdp2}
\alias{Runhdp2}
\title{Run hdp extraction and attribution on a spectra catalog file}
\usage{
Runhdp2(
  input.catalog.file,
  out.dir,
  CPU.cores = 1,
  seedNumber = 1,
  K.guess,
  multi.types = FALSE,
  remove.noise = FALSE,
  test.only = 0,
  overwrite = FALSE,
  verbose = TRUE,
  num.posterior = 4,
  posterior.verbosity = 0,
  cos.merge = 0.9,
  min.sample = 1
)
}
\arguments{
\item{input.catalog.file}{File containing a spectra catalog
in \code{\link[ICAMS]{ICAMS}} format.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{seedNumber}{Specify the random seed for repeatable results.}

\item{K.guess}{Suggested initial value of the number of
signatures, passed to \code{\link[hdp]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, hdp will regard all input spectra as one tumor type.

If \code{TRUE}, hdp will infer tumor types based on the string before "::" in their names.
e.g. tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If \code{multi.types} is a character vector it should be of the same length
as the number of columns in \code{input.catalog}, and each value is the
name of the tumor type of the corresponding column in \code{input.catalog},
e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{remove.noise}{Deprecated; ignored

For result visualization and assessment of \code{hdp} package, select \code{TRUE};
for diagnostic purposes, select \code{FALSE}.}

\item{test.only}{If > 0, only analyze the first \code{test.only} columns
in \code{input.catalog.file}.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}

\item{num.posterior}{Number of posterior sampling chains; can set to
1 for testing.}

\item{posterior.verbosity}{Pass to \code{\link[hdp]{hdp_posterior}}
\code{verbosity}.}

\item{cos.merge}{The cosine similarity threshhold for merging raw clusters
from the posterior sampling chains into "components" i.e. signatures.
This is passed to \code{\link[hdp]{hdp_extract_components}} as
the argument \code{cos.merge}.}

\item{min.sample}{A"components" (i.e. signature) must have at least
this many samples.  This is passed to \code{\link[hdp]{hdp_extract_components}} as
the argument \code{min.sample}.}
}
\value{
The attributed exposure of \code{hdp}, invisibly.
}
\description{
Run hdp extraction and attribution on a spectra catalog file
}
\details{
Creates several
 files in \code{out.dir}. These are:
 TODO(Steve): list the files

 TODO(Wuyang)
}
