% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunmSigActAttributionOnly.R
\name{RunmSigActAttributeOnly}
\alias{RunmSigActAttributeOnly}
\title{Run mSigAct attribution on a spectra catalog file
and known signatures.}
\usage{
RunmSigActAttributeOnly(
  input.catalog,
  gt.sigs.file,
  out.dir,
  CPU.cores = NULL,
  seedNumber = 1,
  test.only = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{input.catalog}{File containing input spectra catalog. Columns are
samples (tumors), rows are mutation types.}

\item{gt.sigs.file}{File containing input mutational signatures. Columns are
signatures, rows are mutation types.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{CPU.cores}{Number of CPUs to use in running
sigfit. For a server, 30 cores would be a good
choice; while for a PC, you may only choose 2-4 cores.
By default (CPU.cores = NULL), the CPU.cores would be equal
to \code{(parallel::detectCores())/2}, total number of CPUs
divided by 2.}

\item{seedNumber}{Specify the pseudo-random seed number
used to run mSigAct. Setting seed can make the
attribution of mSigAct repeatable.
Default: 1.}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.
Default: FALSE}

\item{overwrite}{If TRUE, overwrite existing output.
Default: FALSE}
}
\value{
The attributed exposure of \code{mSigAct}, invisibly.
}
\description{
Run mSigAct attribution on a spectra catalog file
and known signatures.
}
\details{
Creates several
 files in \code{paste0(out.dir, "/sa.output.rdata")}. These are
 TODO(Steve): list the files
}
