% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMutationalPatterns.R
\name{RunMutationalPatterns}
\alias{RunMutationalPatterns}
\title{Run MutationalPatterns extraction and attribution on a spectra catalog file}
\usage{
RunMutationalPatterns(
  input.catalog,
  out.dir,
  CPU.cores = NULL,
  K.exact = NULL,
  K.range = NULL,
  nrun.est.K = 10,
  nrun.extract = 200,
  test.only = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{input.catalog}{File containing input spectra catalog.
Columns are samples (tumors), rows are mutation types.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{CPU.cores}{Number of CPUs to use in running
MutationalPatterns. For a server, 30 cores would be a good
choice; while for a PC, you may only choose 2-4 cores.
By default (CPU.cores = NULL), the CPU.cores would be equal
to \code{(parallel::detectCores())/2}, total number of CPUs
divided by 2.}

\item{K.exact, K.range}{\code{K.exact} is the exact value for
the number of signatures active in spectra (K).
Specify \code{K.exact} if you know exactly how many signatures
are active in the \code{input.catalog}, which is the
\code{ICAMS}-formatted spectra file.

\code{K.range} is A numeric vector \code{(K.min,K.max)}
of length 2 which tell MutationalPatterns to search the best
signature number active in spectra, K, in this range of Ks.
Specify \code{K.range} if you don't know how many signatures
are active in the \code{input.catalog}.

WARNING: You must specify only one of \code{K.exact} or \code{K.range}!

Default: NULL}

\item{nrun.est.K}{Number of NMF runs for each possible number of signature.
This is used in the step to estimate the most plausible number
of signatures in input spectra catalog.}

\item{nrun.extract}{number of NMF runs for extracting signatures and inferring
exposures.}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.
Default: FALSE}

\item{overwrite}{If TRUE, overwrite existing output.
Default: FALSE}
}
\value{
A list contains:\itemize{
\item $signature extracted signatures,
\item $exposure inferred exposures,
} of \code{MutationalPatterns}, invisibly.
}
\description{
WARNING: MutationalPatterns can only do exposure attribution
using SBS96 spectra catalog and signature catalog!
}
\details{
Creates several
 files in \code{out.dir}. These are:
 TODO(Steve): list the files

 TODO(Wuyang)

NOTE: The \code{seed} is hard-coded in
MutationalPatterns as 123456.

\code{pConstant} is harded-coded as 1e-04.
}
