% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunmutSignatures.R
\name{RunmutSignatures}
\alias{RunmutSignatures}
\title{Run mutSignatures extraction and attribution on a spectra catalog file}
\usage{
RunmutSignatures(
  input.catalog,
  out.dir,
  CPU.cores = NULL,
  seedNumber = 12345,
  K.exact = NULL,
  nrun.exact = 1000,
  test.only = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{input.catalog}{File containing input spectra catalog.
Columns are samples (tumors), rows are mutation types.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{CPU.cores}{Number of CPUs to use in running
\code{\link[mutSignatures]{decipherMutationalProcesses}}.}

\item{seedNumber}{Specify the pseudo-random seed number
used to run mutSignatures. Setting seed can make the
attribution of mutSignatures repeatable.
Default: 1.}

\item{K.exact}{\code{K.exact} is the exact value for
the number of signatures active in spectra (K).
Specify \code{K.exact} if you know exactly how many signatures
are active in the \code{input.catalog}, which is the
\code{ICAMS}-formatted spectra file.}

\item{nrun.exact}{number of NMF runs for extracting signatures and inferring
exposures.}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.
Default: FALSE}

\item{overwrite}{If TRUE, overwrite existing output.
Default: FALSE}
}
\value{
The inferred exposure of \code{mutSignatures}, invisibly.
}
\description{
Run mutSignatures extraction and attribution on a spectra catalog file
}
\details{
Creates several
 files in \code{out.dir}. These are:
 TODO(Steve): list the files

 TODO(Wuyang)
}
