% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runsigminer.R
\name{Runsigminer}
\alias{Runsigminer}
\title{Run sigminer extraction and attribution on a spectra catalog file}
\usage{
Runsigminer(
  input.catalog,
  out.dir,
  CPU.cores = NULL,
  seedNumber = 1,
  K.max = NULL,
  test.only = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{input.catalog}{File containing input spectra catalog.
Columns are samples (tumors), rows are mutation types.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{CPU.cores}{Number of CPUs to use in running
sigminer. For a server, 30 cores would be a good
choice; while for a PC, you may only choose 2-4 cores.
By default (CPU.cores = NULL), the CPU.cores would be equal
to \code{(parallel::detectCores())/2}, total number of CPUs
divided by 2.}

\item{seedNumber}{Specify the pseudo-random seed number
used to run SomaticSignatures. Setting seed can make the
attribution of SomaticSignatures repeatable.
Default: 1.}

\item{K.max}{\code{K.max} is the maximum number of signatures
users expect to active in \code{input.catalog}.
As this approach cannot specify \code{K.exact}, you can specify
\code{K.max} = \code{K.exact} If you know exactly how many signatures
are active in the \code{input.catalog}.
On the other hand, you may specify \code{max(K.range)} if you don't
know how many signatures are active in the \code{input.catalog}.}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.
Default: FALSE}

\item{overwrite}{If TRUE, overwrite existing output.
Default: FALSE}
}
\value{
A list contains: \itemize{
\item $signature extracted signatures,
\item $exposure inferred exposures,
} of \code{sigminer}, invisibly.
}
\description{
Run sigminer extraction and attribution on a spectra catalog file
}
\details{
Creates several
 files in \code{out.dir}. These are:
 TODO(Steve): list the files

 TODO(Wuyang)
}
