% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runtcsm.R
\name{Runtcsm}
\alias{Runtcsm}
\title{Run tcsm extraction and attribution on a spectra catalog file}
\usage{
Runtcsm(
  input.catalog,
  out.dir,
  seedNumber = 1,
  K.exact = NULL,
  K.range = NULL,
  covariates = NULL,
  test.only = FALSE,
  overwrite = FALSE,
  feature.file = NULL,
  effect.output.file = NULL,
  sigma.output.file = NULL,
  gamma.output.file = NULL
)
}
\arguments{
\item{input.catalog}{File containing input spectra catalog.
Columns are samples (tumors), rows are mutation types.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{seedNumber}{Specify the pseudo-random seed number
used to run tcsm. Setting seed can make the
attribution of tcsm repeatable.}

\item{K.exact, K.range}{\code{K.exact} is the exact
value for the number of signatures active in spectra (K).
Specify \code{K.exact} if you know exact how many signatures
are active in the \code{input.catalog}, which is the
\code{ICAMS}-formatted spectra file.

\code{K.range} is A numeric vector \code{(K.min,K.max)}
of length 2 which tell tcsm to search the best
signature number active in spectra, K, in this range of Ks.
Specify \code{K.range} if you don't know how many signatures
are active in the \code{input.catalog}.
K.max - K.min >= 3, otherwise an error will be thrown.

WARNING: You must specify only one of \code{K} or \code{K.range}!}

\item{test.only}{If TRUE, only analyze the first 10 columns
read in from \code{input.catalog}.}

\item{overwrite}{If TRUE, overwrite existing output.}
}
\value{
The inferred exposure of \code{tcsm}, invisibly.
}
\description{
Run tcsm extraction and attribution on a spectra catalog file
}
\details{
Creates several
 files in \code{out.dir}. These are:
 TODO(Steve): list the files

 TODO(Wuyang)
}
