% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SummarizeMultiTools}
\alias{SummarizeMultiTools}
\title{Combine \code{\link{SummarizeMultiRuns}} folders from a software package.}
\usage{
SummarizeMultiTools(third.level.dir, tool.dirnames)
}
\arguments{
\item{third.level.dir}{Third level path distinguishing de-novo extraction
+ attribution packages from attribution-only packages.
Examples:
\code{top.dir}/sp.sp/ExtrAttr/
\code{top.dir}/sa.sa/Attr/}

\item{tool.dirnames}{Third level path from the \code{top.dir}. Expected to have
summarized results generated by \code{\link{SummarizeMultiRuns}}.
(multiRun.RDa, ManhattanDist.csv, meanSD.csv, meanSD.Manhattan.dist.csv)
Examples:
\code{"signeR.results"} (Under \code{third.level.dir} "ExtrAttr")
\code{"deconstructSigs.results"} (Under \code{third.level.dir} "Attr")

Here, \code{top.dir} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory within the \code{tool.names}
which stores the software output.}

\item{overwrite}{If TRUE overwrite existing directories and files.}
}
\value{
A list contain c(\code{mean},\code{sd}) of multiple runs:
Cosine similarity
True Positives(TP): Ground-truth signatures which are active in
the spectra, and extracted.
False Negatives(FN): Ground-truth signatures not extracted.
False Positives(FP): Signatures wrongly extracted, not resembling
any ground-truth signatures.
True Positive Rate (TPR, Sensitivity): TP / (TP + FN)
False Discovery Rate (FDR): FP / (FP + TP)
}
\description{
Combine \code{\link{SummarizeMultiRuns}} folders from a software package.
}
