% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SummarizeMultiToolsMultiDatasets}
\alias{SummarizeMultiToolsMultiDatasets}
\title{Combine results for multiple datasets, from different software packages.}
\usage{
SummarizeMultiToolsMultiDatasets(
  dataset.dirs,
  second.third.level.dirname,
  out.dir,
  overwrite = FALSE
)
}
\arguments{
\item{dataset.dirs}{Paths of top-level dataset directories trees you want
to investigate.
E.g. "./S.0.1.Rsq.0.1"}

\item{second.third.level.dirname}{Name of the second.level.dir (e.g. "sp.sp")
and the third.level.dir (e.g. "ExtrAttr") to be investigated.

Examples are: "sp.sp/ExtrAttr", "sa.sa.96/Attr"

Note: \code{multiTools.RDa} are expected to be exist under
\code{dataset.dirs}/\code{second.third.level.dirname}.}

\item{out.dir}{Path of the output directory.}

\item{overwrite}{Whether to overwrite the contents in out.dir if
it already exists. (Default: FALSE)}
}
\description{
Summarize results from each software package in \code{third.level.dir}/\code{tool.dirnames}
(generated by \code{\link{SummarizeMultiRuns}}),
combine them into \code{third.level.dir}.
}
