% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SummarizeMultiToolsOneDataset}
\alias{SummarizeMultiToolsOneDataset}
\title{Combine results for a single dataset, from different software packages.}
\usage{
SummarizeMultiToolsOneDataset(
  third.level.dir,
  toolNames,
  tool.dirnames,
  datasetGroup,
  datasetGroupName,
  datasetSubGroup,
  datasetSubGroupName
)
}
\arguments{
\item{third.level.dir}{Third level path distinguishing de-novo extraction
+ attribution packages from attribution-only packages.
Examples:
\code{top.dir}/sp.sp/ExtrAttr/
\code{top.dir}/sa.sa/Attr/}

\item{toolNames}{Names of software package. (e.g. "sigproextractor")}

\item{tool.dirnames}{Third level path from the \code{top.dir}. Expected to have
summarized results generated by \code{\link{SummarizeMultiRuns}}.
(multiRun.RDa, ManhattanDist.csv, meanSD.csv, meanSD.Manhattan.dist.csv)
Examples:
\code{"signeR.results"} (Under \code{third.level.dir} "ExtrAttr")
\code{"deconstructSigs.results"} (Under \code{third.level.dir} "Attr")

Here, \code{top.dir} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory within the \code{tool.names}
which stores the software output.}

\item{datasetGroup}{Numeric or character vector specifying the group
each dataset belong to.
E.g. For SBS1-SBS5 correlated datasets, we can consider slope as the group:
c("slope=0.1","slope=0.5","slope=1","slope=2","slope=5","slope=10")
Default: "Default"}

\item{datasetGroupName}{Meaning or label of all datasetGroup.
E.g. For SBS1-SBS5 correlated datasets, we can consider \code{"SBS1:SBS5 mutation count ratio"}
as the label of the \code{datasetGroup} slope.}

\item{datasetSubGroup}{Numeric or character vector differentiating
datasets within each group.
E.g. For SBS1-SBS5 correlated datasets, we can consider Pearson's R^2
as the subgroup:
c("Rsq=0.1","Rsq=0.2","Rsq=0.3","Rsq=0.6")
Default: Names of datasets, which are \code{basename(dataset.dirs)}}

\item{datasetSubGroupName}{Meaning or label of all datasetSubGroup.
E.g. For SBS1-SBS5 correlated datasets, we can consider \code{"Pearson's R squared"}
as the label of the \code{datasetSubGroup} Pearson's R^2.}
}
\value{
A list contain c(\code{mean},\code{sd}) of multiple runs:
Cosine similarity
True Positives(TP): Ground-truth signatures which are active in
the spectra, and extracted.
False Negatives(FN): Ground-truth signatures not extracted.
False Positives(FP): Signatures wrongly extracted, not resembling
any ground-truth signatures.
True Positive Rate (TPR, Sensitivity): TP / (TP + FN)
False Discovery Rate (FDR): FP / (FP + TP)
}
\description{
Summarize results from each software package in \code{third.level.dir}/\code{tool.dirnames}
(generated by \code{\link{SummarizeMultiRuns}}),
combine them into \code{third.level.dir}.
}
