% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SummarizeOneToolMultiDatasets}
\alias{SummarizeOneToolMultiDatasets}
\title{Combine results for multiple datasets, from one software tools.}
\usage{
SummarizeOneToolMultiDatasets(dataset.dirs, tool.dirname, out.dir,
  overwrite = FALSE)
}
\arguments{
\item{dataset.dirs}{Paths of top-level dataset directories trees you want
to investigate.
E.g. "./S.0.1.Rsq.0.1"}

\item{tool.dirname}{Name of the second.level.dir (e.g. "sp.sp"),
third.level.dir (e.g. "ExtrAttr") and tool.dir
(e.g. "sigproextractor.results") to be investigated.

One example: "sp.sp/ExtrAttr/sigproextractor.results"

Note: this function expects the summary generated by
\code{SummarizeSigOneSubdir} under \code{dataset.dirs}/\code{tool.dirname}}

\item{out.dir}{Path of the output directory.}

\item{overwrite}{Whether to overwrite the contents in out.dir if
it already exists. (Default: FALSE)}
}
\description{
Summarize results from each software tool in \code{third.level.dir}/\code{tool.dirnames}
(generated by \code{\link{SummarizeMultiRuns}}),
combine them into \code{third.level.dir}.
}
