% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResultsOtherPackages.R
\name{SummarizeSigOneAttrSubdir}
\alias{SummarizeSigOneAttrSubdir}
\title{Assess/evaluate results from packages which can
ONLY do exposure attribution.}
\usage{
SummarizeSigOneAttrSubdir(
  run.dir,
  ground.truth.exposure.dir = paste0(run.dir, "/../../../"),
  overwrite = FALSE
)
}
\arguments{
\item{run.dir}{Lowest level path to results, e.g.
\code{<top.dir>}\code{/sa.sa.96/Attr/YAPSA.results/seed.1/}
Here, \code{<top.dir>} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. For packages which can do both extraction and attribution,
we expect two files, \code{ground.truth.signatures.csv}
and \code{attributed.exposures.csv} are in the folder.}

\item{ground.truth.exposure.dir}{Folder which stores ground-truth exposures.
It defaults to be \code{sub.dir}, i.e. \code{run.dir}/../../}

\item{overwrite}{If TRUE overwrite existing directories and files.}
}
\description{
Packages including but not limited to:
deconstructSigs, YAPSA.
}
\details{
Here, we excluded SignatureEstimation. Although it is also
a package with only attribution, but it has two attribution
algorithms. Therefore the naming of the results are slightly
different from the other two packages.
}
