% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResultshelmsman.R
\name{SummarizeSigOnehelmsmanSubdir}
\alias{SummarizeSigOnehelmsmanSubdir}
\title{Assess/evaluate results from SigProfiler-python (a.k.a. sigproextractor)
Assessment is restricted to v0.0.5.43,
because different version has different folder structure.}
\usage{
SummarizeSigOnehelmsmanSubdir(
  run.dir,
  ground.truth.exposure.dir = paste0(run.dir, "/../../../"),
  overwrite = FALSE,
  hierarchy = FALSE
)
}
\arguments{
\item{run.dir}{Lowest level path to results, e.g.
\code{<top.dir>}\code{/sa.sa.96/ExtrAttr/sigproextractor.results/seed.1/}
Here, \code{<top.dir>} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory
\code{<run.dir>}\code{/SBS96} which
stores SigProfiler results.}

\item{ground.truth.exposure.dir}{Folder which stores ground-truth exposures.
Usually, it refers to \code{sub.dir}, i.e. \code{run.dir}/../../../}

\item{overwrite}{If TRUE overwrite existing directories and files.}

\item{hierarchy}{Whether the user have enabled hierarchy = True
when running sigproextractor.
specifying True or False into sigproextractor will cause the program
to generate different folder structure. (Default: \code{FALSE})}
}
\description{
Assess/evaluate results from SigProfiler-python (a.k.a. sigproextractor)
Assessment is restricted to v0.0.5.43,
because different version has different folder structure.
}
