% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResultsSA.R
\name{SummarizeSigOneSA96Subdir}
\alias{SummarizeSigOneSA96Subdir}
\title{Summarize SNS 96-channel results from SignatureAnalyzer}
\usage{
SummarizeSigOneSA96Subdir(result.dir,
  ground.truth.exposure.dir = paste0(result.dir, "/../../../"),
  which.run = "/best.run/", overwrite = FALSE)
}
\arguments{
\item{result.dir}{Lowest level path to results, that is
\code{<top.dir>}\code{/sa.sa.96/sa.results/},
\code{<top.dir>}\code{/sp.sp/sa.results/},
\code{<top.dir>}\code{/sa.sa.COMPOSITE/sa.results/}, or
\code{<top.dir>}\code{/sp.sa.COMPOSITE/sa.results/}.
Here, \code{<top.dir>} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere.}

\item{ground.truth.exposure.dir}{Folder which stores ground-truth exposures.
It defaults to be \code{sub.dir}, i.e. \code{result.dir}/../../}

\item{which.run}{Name of subdirectory containing the run to summarize.}

\item{overwrite}{If TRUE overwrite existing directories and files.}
}
\description{
Summarize SNS 96-channel results from SignatureAnalyzer
}
\keyword{internal}
