% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResultsSA.R
\name{SummarizeSigOneSASubdir}
\alias{SummarizeSigOneSASubdir}
\title{Summarize results (SBS96, DBS, ID or COMPOSITE) from SignatureAnalyzer}
\usage{
SummarizeSigOneSASubdir(
  run.dir,
  ground.truth.exposure.dir = paste0(run.dir, "/../../"),
  which.run = "/best.run/",
  overwrite = FALSE
)
}
\arguments{
\item{run.dir}{Lowest level path to results, that is
\code{<top.dir>}\code{/sa.sa.96/sa.results/},
\code{<top.dir>}\code{/sp.sp/sa.results/},
\code{<top.dir>}\code{/sa.sa.COMPOSITE/sa.results/}, or
\code{<top.dir>}\code{/sp.sa.COMPOSITE/sa.results/}.

We expect \code{run.dir} contain the best-run subdirectory (e.g. "best.run").
The name of the subdirectory needs to be given to \code{which.run} parameter.

Here, \code{<top.dir>} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere.}

\item{ground.truth.exposure.dir}{Folder which stores ground-truth exposures.
It defaults to be \code{sub.dir}, i.e. \code{run.dir}/../../}

\item{which.run}{Name of subdirectory under \code{run.dir}
containing the run to summarize.}

\item{overwrite}{If TRUE overwrite existing directories and files.}
}
\description{
Summarize results (SBS96, DBS, ID or COMPOSITE) from SignatureAnalyzer
}
\keyword{internal}
