% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResultsSP.R
\name{SummarizeSigOneSigProSSSubdir}
\alias{SummarizeSigOneSigProSSSubdir}
\title{Assess/evaluate results from sigproSS
(a.k.a. SigProfiler Python attribution package)}
\usage{
SummarizeSigOneSigProSSSubdir(
  run.dir,
  ground.truth.exposure.dir = paste0(run.dir, "/../../../"),
  overwrite = FALSE
)
}
\arguments{
\item{run.dir}{Lowest level path to results, e.g.
\code{<top.dir>}\code{/sa.sa.96/ExtrAttr/sigproextractor.results/seed.1/}
Here, \code{<top.dir>} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory
\code{<run.dir>}\code{/SBS96} which
stores SigProfiler results.}

\item{ground.truth.exposure.dir}{TODO(Wu Yang): Fix this
File name which stores ground-truth exposures;
defaults to \code{"ground.truth.syn.exposures.csv"}.
This file can be found in the \code{sub.dir}, i.e. \code{<run.dir>/../../../}}

\item{overwrite}{If TRUE overwrite existing directories and files.}
}
\description{
Assess/evaluate results from sigproSS
(a.k.a. SigProfiler Python attribution package)
}
