% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SummarizeSigOneSubdir}
\alias{SummarizeSigOneSubdir}
\title{Assess/evaluate one result subfolder from a software package}
\usage{
SummarizeSigOneSubdir(
  run.dir,
  ground.truth.exposure.dir,
  extracted.sigs.path,
  attributed.exp.path = NULL,
  overwrite = FALSE,
  summary.folder.name = "summary"
)
}
\arguments{
\item{run.dir}{Lowest level path to result of a run. That is,
\code{top.dir}/sp.sp/ExtrAttr/SignatureAnalyzer.results/seed.1/. or
\code{top.dir}/sa.sa.96/Attr/YAPSA.results/seed.691/
Here, \code{top.dir} refers to a top-level directory which contains the
full information of a synthetic dataset. (e.g. \code{syn.2.7a.7b.abst.v8})
This code depends on a conventional directory structure documented
elsewhere. However there should be a directory within the \code{run.dir}
which stores the software output.}

\item{ground.truth.exposure.dir}{Folder which stores ground-truth exposures.
Usually, it refers to \code{sub.dir}, i.e. \code{run.dir}/../../../}

\item{extracted.sigs.path}{Path to extracted sigs file, e.g.
\code{<run.dir>/SBS96/Selected_Solution/De_Novo_Solution/signatures.PCAWG.format.csv}.}

\item{attributed.exp.path}{Path to attributed exposures file.}

\item{overwrite}{If TRUE overwrite existing directories and files.}

\item{summary.folder.name}{The name of the folder containing summary results.
Usually, it equals to "summary".}
}
\description{
Note: For summarizing sigproextractor or SignatureAnalyzer,
users should use sigproextractor(SigProfiler-Python) v0.0.5.43+
and SignatureAnalyzer 2018-Apr-18.
}
\keyword{internal}
