% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmsmanInteraction.R
\name{ICAMSCatalog2helmsman}
\alias{ICAMSCatalog2helmsman}
\title{Convert Catalogs from ICAMS format to helmsman format}
\usage{
ICAMSCatalog2helmsman(catalog)
}
\arguments{
\item{catalog}{A catalog matrix in ICAMS format. (SNS/DNS/ID)}
}
\value{
a catalog matrix in helmsman format.
}
\description{
Convert Catalogs from ICAMS format to helmsman format
}
