% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmsmanInteraction.R
\name{ReadhelmsmanExposure}
\alias{ReadhelmsmanExposure}
\title{Read Exposure files in helmsman format.}
\usage{
ReadhelmsmanExposure(exposure)
}
\arguments{
\item{exposure}{Exposure file generated by helmsman.
Usually, it is called "W_components.txt".

Return ICAMS/SynSigEval formatted exposure matrix.}
}
\description{
Read Exposure files in helmsman format.
}
