% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runhdp.R
\name{Runhdp}
\alias{Runhdp}
\title{Run hdp extraction and attribution on a spectra catalog file}
\usage{
Runhdp(
  input.catalog,
  out.dir,
  CPU.cores = 1,
  seedNumber = 1,
  K.guess,
  multi.types = FALSE,
  remove.noise = FALSE,
  num.posterior = 4,
  post.burnin = 4000,
  post.n = 50,
  post.space = 50,
  post.cpiter = 3,
  test.only = FALSE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input.catalog}{File containing a spectra catalog
in \code{\link[ICAMS]{ICAMS}} format.}

\item{out.dir}{Directory that will be created for the output;
abort if it already exits.  Log files will be in
\code{paste0(out.dir, "/tmp")}.}

\item{CPU.cores}{Number of CPUs to use in running
\code{\link[hdp]{hdp_posterior}}.}

\item{seedNumber}{Specify the pseudo-random seed number
used to run hdp. Setting seed can make the
attribution of hdp repeatable.
Default: 1.}

\item{K.guess}{Suggested initial value of the number of
signatures, passed to \code{\link[hdp]{dp_activate}} as
\code{initcc}.}

\item{multi.types}{A logical scalar or
a character vector.
If \code{FALSE}, hdp will regard all input spectra as one tumor type,
and will allocate them to one single dirichlet process node.

If \code{TRUE}, hdp will infer tumor types based on the string before "::" in their names.
e.g. Tumor type for "SA.Syn.Ovary-AdenoCA::S.500" would be "SA.Syn.Ovary-AdenoCA"

If it is a character vector, it should be a vector of case-sensitive tumor
types.
e.g. \code{c("SA.Syn.Ovary-AdenoCA", "SA.Syn.Ovary-AdenoCA", "SA.Syn.Kidney-RCC")}.}

\item{remove.noise}{Whether to remove noise signature "hdp.0"? In normal cases scenarios,
only few mutations will be assigned to noise signature.

For result visualization and assessment of \code{hdp} package, select \code{TRUE};
for diagnostic purposes, select \code{FALSE}.}

\item{num.posterior}{Number of posterior sampling chains; can set to
1 for testing.}

\item{post.burnin}{Pass to \code{\link[hdp]{hdp_posterior}}
\code{burnin}.}

\item{post.n}{Pass to \code{\link[hdp]{hdp_posterior}}
\code{n}.}

\item{post.space}{Pass to \code{\link[hdp]{hdp_posterior}}
\code{space}.}

\item{post.cpiter}{Pass to \code{\link[hdp]{hdp_posterior}}
\code{cpiter}.}

\item{test.only}{If TRUE, only analyze the first 10 columns
in \code{input.catalog}.}

\item{overwrite}{If TRUE, overwrite existing output.}

\item{verbose}{If \code{TRUE} then \code{message} progress information.}
}
\value{
The attributed exposure of \code{hdp}, invisibly.
}
\description{
Run hdp extraction and attribution on a spectra catalog file
}
\details{
Creates several
 files in \code{out.dir}. These are:
 TODO(Steve): list the files

 TODO(Wuyang)
}
