% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResults.R
\name{SummarizeOneToolMultiDatasets}
\alias{SummarizeOneToolMultiDatasets}
\title{Combine results for multiple datasets, from one software packages.}
\usage{
SummarizeOneToolMultiDatasets(
  dataset.dirs,
  datasetGroup = NULL,
  datasetGroupName,
  datasetSubGroup = NULL,
  datasetSubGroupName,
  toolName,
  tool.dirname,
  out.dir,
  overwrite = FALSE
)
}
\arguments{
\item{dataset.dirs}{Paths of top-level dataset directories trees you want
to investigate.
E.g. "./S.0.1.Rsq.0.1"}

\item{datasetGroup}{Numeric or character vector specifying the group
each dataset belong to.
E.g. For SBS1-SBS5 correlated datasets, we can consider slope
(SBS1:SBS5 count ratio) as the group:
\code{c(0.1,0.5,1,2,5,10)}
Default: "Default"}

\item{datasetGroupName}{Meaning or label of all datasetGroup.
E.g. For SBS1-SBS5 correlated datasets, we can consider \code{"SBS1:SBS5 mutation count ratio"}
as the label of the \code{datasetGroup} slope.}

\item{datasetSubGroup}{Numeric or character vector differentiating
datasets within each group.
E.g. For SBS1-SBS5 correlated datasets, we can consider Pearson's R^2
as the subgroup:
c(0.1,0.2,0.3,0.6)
Default: Names of datasets, which are \code{basename(dataset.dirs)}}

\item{datasetSubGroupName}{Meaning or label of all datasetSubGroup.
E.g. For SBS1-SBS5 correlated datasets, we can consider \code{"Pearson's R squared"}
as the label of the \code{datasetSubGroup} Pearson's R^2.}

\item{toolName}{Name of software package to be investigated
(e.g. "sigproextractor")}

\item{tool.dirname}{Name of the second.level.dir (e.g. "sp.sp"),
third.level.dir (e.g. "ExtrAttr") and tool.dir
(e.g. "sigproextractor.results") to be investigated.

One example: "sp.sp/ExtrAttr/sigproextractor.results"

Note: this function expects the summary generated by
\code{SummarizeSigOneSubdir} under \code{dataset.dirs}/\code{tool.dirname}}

\item{out.dir}{Path of the output directory.}

\item{overwrite}{Whether to overwrite the contents in out.dir if
it already exists. (Default: FALSE)}
}
\description{
Summarize results from each software package in \code{third.level.dir}/\code{tool.dirnames}
(generated by \code{\link{SummarizeMultiRuns}}),
combine them into \code{third.level.dir}.
}
